Refactor: 채팅기능 구현 끝났습니다. 컨트롤러에는 리퀘스트 파람으로 수정됐는데 테스트 가능한 프론트는 수정중이라서 일단 무더기입니다... 쳇 서비스에 주석처리 된 것들은 알람 관련된 부분이니 이왕이면 건들지 말아주세요...

+++ src/main/java/com/gigajet/mhlb/domain/chat/dto/UnreadMessageCount.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 @Setter
 public class UnreadMessageCount {
     private Long otherUserId;
-    private int unreadCount;
+    private Long unreadCount;
     private String roomId;
     private String type;
